+++ README.md
@@ -1,23 +1,13 @@
-# localstack-aws-sdk-examples
+# 🌟 LocalStack AWS SDK Examples 🌐
 
-This repository contains examples that demonstrate how to integrate AWS SDKs for different languages with LocalStack.
+This repository showcases how to integrate AWS SDKs with LocalStack across various programming languages. 🚀
 
-* JavaScript
-  * [/javascript/amplify-js](/javascript/amplify-js): [amplify-js-local](https://github.com/localstack/amplify-js-local)
-* Go
-  * [/go/s3-presign-v1.go](/go/s3-presign-v1.go): [aws-sdk-go (v1)](https://github.com/aws/aws-sdk-go)
-  * [/go/s3-basic-v2.go](/go/s3-basic-v2.go): [aws-sdk-go (v2)](https://github.com/aws/aws-sdk-go-v2)
-* PHP
-  * [/php](/php): [aws-sdk-php](https://github.com/aws/aws-sdk-php)
-* Java: [/java/java-sdk-examples](/java/java-sdk-examples):
-  * S3 v1: [java/java-sdk-examples/java-sdk-v1/src/main/java/v1/s3](java/java-sdk-examples/java-sdk-v1/src/main/java/v1/s3)
-  * S3 v2: [java/java-sdk-examples/java-sdk-v2/src/main/java/v2/s3](java/java-sdk-examples/java-sdk-v2/src/main/java/v2/s3)
-  * DynamoDB v1: [java/java-sdk-examples/java-sdk-v1/src/main/java/v1/dynamodb](java/java-sdk-examples/java-sdk-v1/src/main/java/v1/dynamodb)
-  * DynamoDB v2: [java/java-sdk-examples/java-sdk-v2/src/main/java/v2/dynamodb](java/java-sdk-examples/java-sdk-v2/src/main/java/v2/dynamodb)
+## 🚀 Running the Examples
 
+Each folder is self-contained and can be independently run using `Makefile`. To execute an example, simply use the `make run` command.
 
-## Downloading a single folder
+## 📥 Downloading a Single Folder
 
-1. In the root of the repo press `.` or replace `.com` with `.dev` in URL to open the repo in GitHub's internal editor.
-2. In the explorer pane (left side or press Ctrl+Shift+E), right-click on the required file/folder and select download.
-3. In the dialog box, choose the directory on your disk under which you want the selected file/folder to be saved.
+1. Open the repo in GitHub's internal editor by pressing `.` at the repo root or changing `.com` to `.dev` in the URL.
+2. In the explorer pane (left side or press Ctrl+Shift+E), right-click on the desired file/folder and select 'Download'.
+3. In the dialog box, choose where to save the selected file/folder on your disk.
