+++ go/s3/presign/v1/Makefile
@@ -1,0 +1,2 @@
+run:
+	go run main.go
