+++ go/s3/presign/v1/go.mod
@@ -1,0 +1,7 @@
+module github.com/localstack/localstack-aws-sdk-examples
+
+go 1.20
+
+require github.com/aws/aws-sdk-go v1.48.3
+
+require github.com/jmespath/go-jmespath v0.4.0 // indirect
