+++ go/s3/s3-crud-sdk-v2/Makefile
@@ -1,0 +1,2 @@
+run:
+	go run main.go
